resolves merge conflicts in server.js
