+++ app.js
@@ -35,7 +35,7 @@ db
         displayName: Sequelize.STRING,
       });
       exports.Student.sync() //syncs the model with the database. This creates the Students table.
-      exports.Teacher.sync()
+      exports.Teacher.sync() //syncs the model with the database. This creates the Teacher table.
     }
   })
