+++ server.js
@@ -3,6 +3,7 @@ var app = express();
 var server = require('http').Server(app);
 var io = require('socket.io')(server);
 var teachers = require('./config.js').teachers;
+var utils = require('./utils.js');
 
 ////////////////////////////////////////////////////
 var request = require('request');//simplified HTTP request client
@@ -85,11 +86,11 @@ function(req, res){
 app.get('/github/callback',
   passport.authenticate('github', { failureRedirect: '/' }),
   function(req, res) {
-    if (isTeacher(req.user.displayName, teachers)) {
-      createUser(req.user.displayName, db.Teacher);
+    if (utils.isTeacher(req.user.displayName, teachers)) {
+      utils.createUser(req.user.displayName, db.Teacher);  
       res.redirect('/teacher');
     } else {
-      createUser(req.user.displayName, db.Student);
+      utils.createUser(req.user.displayName, db.Student);
       res.redirect('/student/*');
     }
 
@@ -175,54 +176,5 @@ io.on('connection', function (socket) {
     io.sockets.emit('resolved');
   });
 });
-
-
-//////////////////HELPER FUNCTIONS///////////////////////
-
-var isTeacher = function(displayName, teachers){
-  for (var teacher in teachers) {
-    if (teachers[teacher] === displayName){
-      return true;
-    }
-  }
-  return false;
-};
-
-var createUser = function(displayName, model){
-  console.log('////////MODEL', model);
-  model
-  .create({
-    displayName: displayName
-  })
-  .complete(function(err, user) {
-    if(err){
-      console.log('error: ' + err)
-    } else{
-      console.log('user is saved! ' + user)
-    }
-  })
-};
-
-var getUsersEmails = function(username, accessToken) {
-  var options = {
-    headers: {
-      'User-Agent':    'JavaScript.ru',
-      'Authorization': 'token ' + accessToken
-    },
-
-    json:    true,
-    url:  'https://api.github.com/users:' + username //  'https://api.github.com/user/emails'
-  };
-
-  request(options, function (error, response, body) {
-    if (!error && response.statusCode == 200) {
-      console.log(response);
-    }
-    console.log(response.statusCode)
-  })
-};
-
-
-
 
 }
